requirements.txt
Flask
Flask-CORS
gunicorn
pymongo[srv]
dnspython
google-genai
python-docx
unidecode
langdetect
cloudinary
requests
werkzeug
pytest
python-dotenv
